Added the seeder to incorporate a few example contacts and to make sure that everything is linked together properly.

+++ contacts_migration.php
@@ -9,7 +9,7 @@ $createContactsTable = 'CREATE TABLE IF NOT EXISTS contacts (
 	id INT UNSIGNED NOT NULL AUTO_INCREMENT,
 	first_name varchar(200) NOT NULL,
 	last_name varchar(200) NOT NULL,
-	phone_number INT NOT NULL,
+	phone_number DOUBLE NOT NULL,
 	company varchar(200),
 	email varchar(100),
 	PRIMARY KEY (id)
